added better error printing for users with miette and refactored/restructured code
