+++ Cargo.toml
@@ -19,8 +19,10 @@ license-file = ["LICENSE", "14"]
 chrono = "0.4.41"
 clap = { version = "4.5.41", features = ["derive"] }
 crossterm = { version = "0.29.0", features = ["osc52", "event-stream"] }
+miette = { version = "7.6.0", features = ["fancy"] }
 serde = { version = "1.0.219", features = ["derive"] }
 serial2-tokio = { version = "0.1.15", features = ["windows"] }
+thiserror = "2.0.16"
 tokio = { version = "1.47.0", features = ["full"] }
 toml = "0.9.5"
 # tracing = "0.1"
